+++ scripts/run_custom_classifier.py
@@ -22,11 +22,11 @@ import json
 from docopt import docopt
 
 import sense.display
-from sense import engine
 from sense import feature_extractors
 from sense.controller import Controller
 from sense.downstream_tasks.nn_utils import LogisticRegression
 from sense.downstream_tasks.nn_utils import Pipe
+from sense.downstream_tasks.nn_utils import load_weights
 from sense.downstream_tasks.postprocess import PostprocessClassificationOutput
 
 
@@ -42,10 +42,10 @@ if __name__ == "__main__":
 
     # Load original feature extractor
     feature_extractor = feature_extractors.StridedInflatedEfficientNet()
-    checkpoint = engine.load_weights('resources/backbone/strided_inflated_efficientnet.ckpt')
+    checkpoint = load_weights('resources/backbone/strided_inflated_efficientnet.ckpt')
 
     # Load custom classifier
-    checkpoint_classifier = engine.load_weights(os.path.join(custom_classifier, 'classifier.checkpoint'))
+    checkpoint_classifier = load_weights(os.path.join(custom_classifier, 'classifier.checkpoint'))
     # Update original weights in case some intermediate layers have been finetuned
     name_finetuned_layers = set(checkpoint.keys()).intersection(checkpoint_classifier.keys())
     for key in name_finetuned_layers:
