+++ scripts/run_fitness_rep_counter.py
@@ -18,12 +18,12 @@ Options:
 from docopt import docopt
 
 import sense.display
-from sense import engine
 from sense import feature_extractors
 from sense.controller import Controller
 from sense.downstream_tasks.fitness_rep_counting import INT2LAB
 from sense.downstream_tasks.nn_utils import LogisticRegression
 from sense.downstream_tasks.nn_utils import Pipe
+from sense.downstream_tasks.nn_utils import load_weights
 from sense.downstream_tasks.postprocess import PostprocessClassificationOutput
 from sense.downstream_tasks.postprocess import PostprocessRepCounts
 
@@ -45,7 +45,7 @@ if __name__ == "__main__":
     # Load a logistic regression classifier
     gesture_classifier = LogisticRegression(num_in=feature_extractor.feature_dim,
                                             num_out=5)
-    checkpoint = engine.load_weights('resources/fitness_rep_counting/efficientnet_logistic_regression.ckpt')
+    checkpoint = load_weights('resources/fitness_rep_counting/efficientnet_logistic_regression.ckpt')
     gesture_classifier.load_state_dict(checkpoint)
     gesture_classifier.eval()
