+++ scripts/run_fitness_tracker.py
@@ -24,13 +24,13 @@ import torch
 from docopt import docopt
 
 import sense.display
-from sense import engine
 from sense import feature_extractors
 from sense.controller import Controller
 from sense.downstream_tasks import calorie_estimation
 from sense.downstream_tasks.fitness_activity_recognition import INT2LAB
 from sense.downstream_tasks.nn_utils import LogisticRegression
 from sense.downstream_tasks.nn_utils import Pipe
+from sense.downstream_tasks.nn_utils import load_weights
 from sense.downstream_tasks.postprocess import PostprocessClassificationOutput
 
 
@@ -55,7 +55,7 @@ if __name__ == "__main__":
     # Load fitness activity classifier
     gesture_classifier = LogisticRegression(num_in=feature_extractor.feature_dim,
                                             num_out=81)
-    checkpoint = engine.load_weights('resources/fitness_activity_recognition/mobilenet_logistic_regression.ckpt')
+    checkpoint = load_weights('resources/fitness_activity_recognition/mobilenet_logistic_regression.ckpt')
     gesture_classifier.load_state_dict(checkpoint)
     gesture_classifier.eval()
