+++ scripts/run_gesture_recognition.py
@@ -18,12 +18,12 @@ Options:
 from docopt import docopt
 
 import sense.display
-from sense import engine
 from sense import feature_extractors
 from sense.controller import Controller
 from sense.downstream_tasks.gesture_recognition import INT2LAB
 from sense.downstream_tasks.nn_utils import LogisticRegression
 from sense.downstream_tasks.nn_utils import Pipe
+from sense.downstream_tasks.nn_utils import load_weights
 from sense.downstream_tasks.postprocess import PostprocessClassificationOutput
 
 
@@ -44,7 +44,7 @@ if __name__ == "__main__":
     # Load a logistic regression classifier
     gesture_classifier = LogisticRegression(num_in=feature_extractor.feature_dim,
                                             num_out=30)
-    checkpoint = engine.load_weights('resources/gesture_detection/efficientnet_logistic_regression.ckpt')
+    checkpoint = load_weights('resources/gesture_detection/efficientnet_logistic_regression.ckpt')
     gesture_classifier.load_state_dict(checkpoint)
     gesture_classifier.eval()
