+++ sense/engine.py
@@ -1,7 +1,6 @@
 import numpy as np
 import queue
 import torch
-import os
 
 from threading import Thread
 from typing import List
@@ -146,17 +145,3 @@ class InferenceEngine(Thread):
 
         return predictions
 
-
-def load_weights(checkpoint_path: str):
-    """
-    Load weights from a checkpoint file.
-
-    :param checkpoint_path:
-        A string representing the absolute/relative path to the checkpoint file.
-    """
-    try:
-        return torch.load(checkpoint_path, map_location='cpu')
-    except:
-        raise Exception('ERROR - Weights file missing: {}. To download, please go to '
-                        'https://20bn.com/licensing/sdk/evaluation and follow the '
-                        'instructions.'.format(checkpoint_path))
